Set example params in change nodes
Previously, the parameters used in TIE requests in the examples could be
set across multiple nodes within the flow - the starting inject node, a
change node, and/or the request node itself. In this commit, the
examples have been updated to uniformly set all of the available TIE
request parameters from within the change node which immediately
precedes the TIE request node in the flow.

This commit also updates links to the TIE client documentation to
reference the JavaScript SDK doc pages instead of the Python SDK doc
pages and fixes cases where the type of the `publicKeySha1` property was
listed as `number` to instead be the correct type, `string`.